# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.User:
    description: 用户信息模型
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      last_login:
        example: "2024-01-01T00:00:00Z"
        type: string
      role:
        enum:
        - user
        - admin
        example: user
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      username:
        example: john_doe
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - role
    - username
    type: object
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Fiber框架模板API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Template API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: 获取系统中所有用户的列表
      produces:
      - application/json
      responses:
        "200":
          description: 包含用户列表的响应
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有用户
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建一个新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功的响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求体错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 通过ID删除特定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功的响应
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 通过ID获取特定用户的详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 包含用户信息的响应
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 用户未找到
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 通过ID更新特定用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功的响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求体错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新用户
      tags:
      - 用户管理
schemes:
- http
swagger: "2.0"